feat: 회원가입 로직 수정 - memeberDetail DTO 수정 (학과랑 학교 삭제) - pref까지 받고 회원가입 - pref post 삭제
